requirements.txt
astroid==2.5.6
click==7.1.2
dnspython==2.1.0
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pydantic==1.8.2
pylint==2.8.2
pymongo==3.11.4
python-dotenv==0.17.1
PyYAML==5.4.1
starlette==0.14.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wrapt==1.12.1
